+++ src/main/java/de/code_notes/backend/services/NoteService.java
@@ -1,12 +1,18 @@
 package de.code_notes.backend.services;
 
+import static org.springframework.http.HttpStatus.BAD_REQUEST;
+import static org.springframework.http.HttpStatus.CONFLICT;
+
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.validation.Validator;
+import org.springframework.web.server.ResponseStatusException;
 
 import de.code_notes.backend.entities.Note;
 import de.code_notes.backend.repositories.NoteRepository;
+import jakarta.annotation.Nullable;
 import lombok.extern.log4j.Log4j2;
 
 
@@ -21,10 +27,16 @@ public class NoteService {
     private TagService tagService;
 
     @Autowired
+    private AppUserService appUserService;
+
+    @Autowired
     private CodeBlockService codeBlockService;
     
     @Autowired
     private PlainTextBlockService plainTextBlockService;
+
+    @Autowired
+    private Validator validator;
 
 
     /**
@@ -45,22 +57,35 @@ public class NoteService {
      */
     public Note save(Note note) {
 
-        // get note from db
-        Note oldNote = getById(note.getId());
-        
+        // case: falsy param
+        if (note == null)
+            throw new IllegalStateException("Failed to save note. 'note' cannot be null");
+
+        // validate
+        validateAndThrow(note);
+
+        // case: appUser does not exist
+        if (!this.appUserService.exists(note.getAppUser()))
+            throw new ResponseStatusException(CONFLICT, "Failed to save note. 'note.appUser' does not exist");
+
         // case: note exists
+        Note oldNote = getById(note.getId());
+        // TODO: test
         if (oldNote != null) {
             // delete old blocks
             this.codeBlockService.deleteAll(oldNote.getCodeBlocks());
             this.plainTextBlockService.deleteAll(oldNote.getPlainTextBlocks());
         }
 
-        // save tags and update set
-        note.setTags(this.tagService.saveOrGetNoteTags(note));
+        // save and get tags
+        if (note.getTags() != null)
+            note.setTags(this.tagService.saveOrGetNoteTags(note));
 
         note = this.noteRepository.save(note);
 
-        this.tagService.removeOrphanTags();
+        // remove tags that are no longer related to any note
+        // TODO: test
+        this.tagService.removeOrphanTags(note.getAppUser());
 
         return note;
     }
@@ -70,7 +95,7 @@ public class NoteService {
      * @param id
      * @return note with given id or {@code null}
      */
-    public Note getById(Long id) {
+    public Note getById(@Nullable Long id) {
 
         if (id == null)
             return null;
@@ -84,7 +109,7 @@ public class NoteService {
      * 
      * @param note
      */
-    public void delete(Note note) {
+    public void delete(@Nullable Note note) {
 
         if (note != null)
             this.noteRepository.delete(note);
@@ -96,8 +121,21 @@ public class NoteService {
      * 
      * @param id
      */
-    public void delete(Long id) {
+    public void delete(@Nullable Long id) {
 
         this.delete(getById(id));
+    }
+
+
+    // TODO: put this inside abstract service
+    private boolean validateAndThrow(Note note) {
+
+        if (note == null)
+            throw new IllegalStateException("Failed to validate note. 'note' cannot be null");
+
+        this.validator.validateObject(note)
+                      .failOnError((message) -> new ResponseStatusException(BAD_REQUEST, "'note' invalid"));
+
+        return true;
     }
 }
